+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Changed
     - Refactor vue3app ``fetch()`` calls (#2475)
     - Refactor vue3app ``notifyCb`` access (#2470)
     - Refactor vue3app ``gridUtils`` storage and params (#2448)
+    - Update vue3app new row save enabling (#2490)
 
 Fixed
 -----